Fix jittery BottomSheet behavior with new Support library.
In our customized BottomSheet, we attempt to give it a maximum width,
which is fine, but for the height parameter we were specifying
wrap_content, which apparently causes a jittery behavior in the newest
Support library. Setting it to match_parent solves this issue.

Additionally, the "overlay" button used in the link preview dialog should
no longer latch on to the R.id.content view, but rather R.id.container,
which seems to be another new thing introduced in the latest library.

Change-Id: I4bd0603487a48b304c00200552d3300124a369c7