Utterly standardize button styles throughout the app.
Sorry, this had to be done sooner than later. We have a whole lot of
inconsistency in all the different buttons in our layouts. Furthermore we
also use an outdated and convoluted system of selectors and shapes to
represent the backgrounds of some of the buttons.  This patch does all of
the following:

- Create a canonical App.Button style that is the base for all buttons in
  the app, which inherits from the baseline Material button.
- Create a couple more children of that style, with different colors.
- The colors of the buttons are now represented by simple color state
  resources. All of the old selectors and shapes are now gone.
- All TextViews that are used as buttons are now AppCompatTextViews, which
  is more correct for utilizing properties like app:backgroundTint.

From this time forward, if we need to put a button in a layout, it shall
be an AppCompatTextView, with a style of "App.Button" or one of the color
derivatives.

This will have some slight visual changes for the buttons in the Login and
CreateAccount activities, since those buttons weren't really standard, but
I believe these changes are for the better.

Change-Id: I4389dbed154fdb649f94be2ccd93ef901d77c8d4