Fix possible crash when highlighting text on API 23.
https://rink.hockeyapp.net/manage/apps/226649/app_versions/204/crash_reasons/173095781

This occurs only on Android 6.0 (API 23):
To reproduce:
Find any TextView that has text selection enabled, and that has a
LinkMovementMethod associated with it. Then long-press to select some text
inside it, and then click away from the selection (on another part of the
same TextView). It crashes with an IllegalArgumentException.

This workaround patches the internals of the TextView itself. We already
have a "custom" patched TextView component, which is AppTextView, so we
can apply the fix there, and just switch all instances of TextViews that
have selection enabled to become AppTextViews.

Change-Id: If2f24c4cd3a799c5ecd2e32d29372ba9b2c1fd83