Add a setting for app color theme
This adds a setting in the main preferences menu to toggle the color theme
between light and dark.

There's a wrinkle!  A SwitchPreferenceCompat is meant to deal with boolean
preference values, and the app theme is stored as an integer (0 for light
and 1 for dark).

So, I've created an abstract IntSwitchPreferenceCompat class that
translates between checked/unchecked states and a pair of integer
values, and a ThemeSwitchPreferenceCompat concrete subclass that wires
it for the dark mode value when checked and light when unchecked.

Bug: T160411
Change-Id: I38e38d9e8b99d28dc4cdb3bd7cee71764dd1e55d