When saving a page, persist its thumbnail in reading list db row.
The predecing patches correctly saved the thumbnail to local storage, but
didn't actually persist the thumbnail URL in the metadata of the reading
list row in the database.  This patch adds in the thumbnail URL and
description to the db row of the page that is being saved, so that they're
shown correctly when viewing reading lists offline.

Bug: T166409
Change-Id: I92f882c42449a7ef17188aa9e0ade16f296ef344