Fix playback of audio files.
Here's where my research on this has led:
(see previous patch set for earlier notes)

To debug this more definitively, I subclassed InputStream and overrode all
of its methods, and added log messages to see which methods were being
called after the InputStream gets passed back to the WebView for
processing.  What I found is very interesting:

The first thing that gets called is the available() method.  But wait...
the byteStream given by OkHttp returns zero when calling available()! (or
rather, available() returns 0 when it's called before any read() methods.)

This seems to be in violation of the platform's InputStream documentation,
which says that the available() return value is "an estimate of the number
of bytes that can be read (or skipped over) from this input stream without
blocking, or 0 when it reaches the end of the input stream."

The above logic implies that any consumer of OkHttp's byteStream that
relies on available() is bound to fail.  And it's looking like the
consumer that plays back audio files does indeed rely on available() to
give it the total number of bytes left in the stream to read.

The solution proposed here is to subclass InputStream, initialize it with
the total byte count that the stream will contain, and explicitly keep
track of the byte count with every read() and skip() call, so that any
call to the available() method will produce a meaningful value.

Of course, we will restrict usage of this subclass to the specific mime
type of "application/ogg" for now.

Bug: T164520
Change-Id: I1aa424d43ac4dcfbb18787da41f53e0a55ac9b0a