Fix possible crash(es) when saving images from Gallery or Feed.
Half-hygiene, half crash fix:
This patch brings our entire logic of saving images more up-to-date with
our current patterns and practices. It removes the necessity to pass and
persist Activity and Context instances in several places, and introduces
proper Callback behavior that we know and love.

It should also fix the following crash, if not a few others:
https://rink.hockeyapp.net/manage/apps/226649/app_versions/204/crash_reasons/173113217

Change-Id: Ifc04526b8ce6107ee190c41644fb8a24c91d282c