Hide reading list title when searching with no results.
This is a tiny design tweak, to comply with the design recommendation in
the task.  The only thing that can't be accomplished at the moment is
vertical centering of the no-results view. This doesn't seem to be very
easy to do with a CoordinatorLayout + NestedScrollView, and we can
investigate it at a later time.

Bug: T156086
Change-Id: Idfc821a628ae4100b86893e3d9993300146a67bf