Fix proper retrieval of User Options.
Here's the thing:
When we make the API call to get the "options" for the current user, the
API actually always returns success, regardless of whether the user is
logged in or not:

- If the user is logged in, the API returns the options for that user.
- If the user is not logged in, the API returns the options for the
  *anonymous* user implicitly assigned to that IP address!

This is a problem for us because there are some cases where the app
"believes" the user is logged in, but the API disagrees. This can happen
for several reasons, including if the OAuth cookie is expired, or if the
app data is cleared, but the user's account is still present in the
AccountManager.

Anyway, the place where this presents an actual problem is when we
retrieve the user options. In other parts of the code where we rely on the
user being logged in, we depend on CsrfTokenClient to implicitly log in
the user before proceeding.  So, this patch modifies the get() method that
retrieves user options, so that it goes through the CsrfTokenClient cycle
before getting the options.  It won't actually use the token, but this
will ensure that the user is properly logged in.

Change-Id: I5820ea94549fdcc2e88a3d6804d2689c2cb9943e