Add bulk reading list actions, p. 1
Puts the basic infrastructure for bulk actions in place and enables
bulk saving for/unsaving from offline.

Bulk adding to another list is a little more involved and is WIP. I'll
submit that later. But since I'm out soon, and in the interest of keeping
patches manageable in size, I thought I'd submit this now.

Bug: T160637
Change-Id: I94db41928afbf86adcb47e7af8133905ac4813eb