Lint: Downgrade ExtraTranslation from error to warning
The ExtraTranslation rule checks for strings for which there are
translations but no default.  This is expected when removing strings
(according to Siebrand, the translations will be removed from TWN with no
further intervention needed from us) but the Android linter throws an
error by default for these.  Let's downgrade it to a warning.

Change-Id: Iff1df045ece3669bf177ff1d8424d9fc53036750