ZIM Compilations: the beginning.
This patch forms the kernel of how the rest of the app will interface with
ZIM compilations. We introduce the OfflineManager class, which is a
singleton that will provide content from a compilation. Whether it's using
one or more compilations simultaneously is completely transparent.

This does not yet introduce any actual usage of compilations in the app,
but simply brings the foundation for upcoming patches.

Also, tests.

Bug: T163584
Bug: T166562
Change-Id: If3eac3ce42412d605d51ea26c6232d8225d6a8c7