Introduce a disk usage component, and use it for Compilations.
The only difficulty encountered in the making of this component is that it
doesn't seem to be possible for us to determine how much space our app
uses. The system doesn't provide this information easily.

Instead this component currently only states the amount of space used by
compilations, whose magnitude will likely overshadow the storage footprint
of the app itself.

Bug: T164387
Change-Id: Ia97c84cf6c43d9fdf4e4b47513fbc972ea32382b