Update Fresco to 1.4.0
Release notes:
https://github.com/facebook/fresco/releases/tag/v1.4.0

A few breaking changes are listed, but none that seem to apply to us.
They're also claiming improved performance with animated images. And most
importantly (although it's not listed in the release notes) it contains a
fix for inter-activity transitions, which weren't working ideally on API
25. Stay tuned.

Also updated ProGuard based on their guidelines.

Change-Id: Ia074579f986501320218cbfe9e68de960f383f8e