Fix: Crash when opening settings
A recent change introduced a crash when opening the app settings. This
was because of a newly added v21 AppTheme style which was intended to
supplement the main AppTheme style but instead was overriding it. The
v21 AppTheme lacked a preferenceTheme attribute, which would cause the
app to crash with an IllegalStateException when opening the settings.

Change-Id: I29680bee09021d9d4de6277dfcd72751ba56f67f