Fix infinite theme change loop in PageActivity.
When changing the theme in PageActivity (while online) it enters an
infinite loop of light/dark theme changes.  This is because of a race
condition between the UserOption uploader uploading the new value and the
UserOption *downloader* downloading the previous value and triggering a
theme change to the old value.

This patch requests a sync of UserOptions only when the activity is
launched freshly (i.e. not restarted via theme change).

Bug: T170756
Change-Id: I830aef92c1c57dc64e42c49550a0b2fcb402534b