Update permissions logic for compilations.
This restricts the request for Read permissions to the CompilationActivity,
and removes the request from the Feed and from PageActivity.

Note: this leaves the framework for asking for the Read permission in
BaseActivity, in case we need to bring back the automatic permission logic
when launching other activities.

This also detects whether the user selected "Never ask again" in the
permission dialog, by making use of the
shouldShowRequestPermissionRationale() method, which returns false if the
user selects "Never ask again."  Unfortunately this method also returns
false if we've never asked for the permission before.  Therefore we need
to persist an additional variable ourselves that tracks whether we've
asked for the permission at least once.

Change-Id: I57012d752df0c4f55106487a37d4516a10252091