Fix: Use groups, not implicitgroups, to determine user rights.
The groups associated with a user account determine the user's rights to
perform actions such as editing a semi-protected or protected article.
These can be obtained from the 'groups' property for a user via the MW
API.  'Implicitgroups' is a subset of 'groups' consisting of only those
groups that a user is added to automatically, such as 'autoconfirmed.'
For many users 'groups' will be identical to 'implicitgroups' but this is
not always the case.  For instance, for my staff account, which has sysop
rights on testwiki:

https://phabricator.wikimedia.org/P5792

We should look to 'groups' and not 'implicitgroups' to determine the
correct set of groups assigned to the user account.

Change-Id: Ib815d4b8de7c34ea4c27045d9297aba6981936a7