Compilations: close the loop.
This implements downloading of compilations.

Some architecture notes:
This introduces an abstract Fragment that becomes an observer of the
system DownloadManager, by polling it at a predefined interval.

So then, the other fragments that show lists of compilations (or a single
compilation) will automatically become subscribers to the polling of the
DownloadManager.  Upon every poll event, we'll update each compilation
item with the appropriate state that the DownloadManager reports.

If a compilation item is in the middle of being downloaded, we show the
DownloadControlView component inline with the item.

TODO:
- Implement deleting a downloaded compilation from disk.
- Implement more robust behavior upon successful download completion.

Bug: T167523
Bug: T164764
Bug: T167527
Change-Id: I2741c14894d990c59c30bd685546c2dd61c5cab3