Fake out Explore toolbar overlap as one single bar.
This will dynamically set or remove the elevation on the Toolbar in the
Feed, based on the scroll state of the feed contents (the scroll state is
borrowed from whether the Search icon is currently shown), thereby making
it look like one continuous app bar when scrolled to the top.

The alternative (using a CollapsingToolbarLayout with AppBarLayout with
SwipeRefreshLayout with custom Toolbar with custom overlapping scroll
behavior) is significantly more complex.

Bug: T170831
Change-Id: I73390ba0dd41690c1289ae76388a2717ebc5d30a