Update testing scripts to accept Android SDK location argument
Updates our non-periodic testing job scripts to accept an optional
command line argument containing the location of an Android SDK
installation.  This can be configured at the job level in Jenkins.
(The jobs have been updated accordingly in Jenkins via the web UI but
this change should be memorialized in the integration/config repo.)

Bug: T171811
Change-Id: If5801d0556e6d084e236136e9b2957a7cef32538