Update FileUtil read* methods
The current Scanner-based implementation of readFile fails
persistently but for unknown reasons in Docker, throwing a
NoSuchElementException for some files.

This replaces both Scanner-based read* implementations with calls to
Commons I/O library functions, which should be more robust (and indeed
cause no trouble in Docker).

Commons I/O is added as a testing dependency only in order to avoid
increasing the APK size.

Change-Id: I6d034957c7a598d68de362394d21d5153b98d52f