Shorten the rationale
The rationale shown when storage permission is revoked seems a bit
too long and makes the assumption that the user would see the dialog
to grant permission each time he does the concerned action. It's not
always the fact as the user has the choice of "Don't show again" and
may have used it to avoid the dialog from popping up.

Shorten the message thus making no assumptions about the dialog being
shown.

Change-Id: Icd554217c9cf3cf955e4ddfe7c3ce2ba24d9a1a8