Remove incorrect RESTBase error handling
Several RESTBase response model classes currently expect an error response
object to be contained in an "error" field similar to the MediaWiki API's
error response structure, but RESTBase doesn't return errors this way; it
simply sends the error object.  Hence it will never be picked up in the
way it's currently expected.

This removes the incorrect code described above.  In the error case, an
HttpStatusException will be thrown, which will help us diagnose the error.
The response body can be examined through interactive debugging if needed.

Bug: T150859
Change-Id: I154fa0df9f7cc4b9dca307b6bf5497993babc50c