Fix saving pages for offline use (convincingly).
An enormous thanks to Rita for continuing to test offline reading lists,
and managing to hit upon a reproducible failure mode.

The failure mode: Disable RestBase in settings!
That's right. This means that the offline feature hasn't been working for
any users in China, or for any other user for whom the fallback mechanism
was triggered from Rb to Mw, for whatever reason.

The bug is that the SavedPageSyncService has a line of code where it
baseically expects the response to be a RestBase reply, instead of a
MobileView reply.  This has now been fixed.

So... when can we officially deprecate support for MobileView, again? ;)

Bug: T171939
Bug: T172900
Bug: T169819
Change-Id: I4b888923de48affa837d6c33fb1eae1a74738ff9