Fix: ellipsize reading list descriptions
Reading list descriptions were not being ellipsized despite our specifying
maxLines="2" and ellipsize="end" in the layout xml.  Apparently because of
a platform bug, tt turns out that the ellipsize directive doesn't work for
TextViews with a TextWatcher set on them, and therefore won't work with a
GoneIfEmptyTextView.  We don't need to use that class anyway, so this
swaps in a plain TextView.

Added some cautionary comments to GoneIfEmptyTextView to reflect this.

Bug: T164926
Change-Id: Ie97a9407e83c10719dbcbc6c0554631a7e30b034