Fix shifting-up of lead paragraph when not using RestBase.
When RestBase is disabled, I noticed that the lead paragraph was no longer
being shifted to the top (i.e. the infobox appears at the top instead).
This is because the recent lazy-loading functionality creates a virtual
DOM, whereas the shifting transform relies on the lead section to be
inside a real DOM for measuring the physical height of the lead paragraph.

This patch moves the transform logic for the lead section to a point after
the section is appended to the real DOM.

Change-Id: I100dea60b46cc20f71da0c2d8e296319a90a7616