Prevent possible memory leak in OfflineManager.
When searchForCompilations() is invoked, the Callback that is passed into
it can be an anonymous class belonging to an Activity.  When the
searchTask completes, it can technically hold on to a reference of
Callback indefinitely.

This patch explicitly sets searchTask to null as soon as the task itself
finishes, so that any references to the Callback (and thus references to
an Activity) are not held for longer than necessary.

Change-Id: I197dbabadae59bd93d1ac6e3adf49a2e4fd76acc