Log HTTP error status codes for page save failures
We are attempting to log errors when saving pages in HockeyApp in attempt
to debug remaining page save issues, but the logged errors currently
contain no useful information; the most we can deduce is that the server
returned an error response code.  This updates the HttpStatusException to
provide the response code via getMessage for remote logging.

Change-Id: I9f45f887c01245a7fa26b128268c7623339682f1