Make DownloadManager notification clickable.
This creates an app-wide receiver for DownloadManager events, and watches
for clicks of the DownloadManager notification. If one of the files being
downloaded is a compilation, we launch the LocalCompilations activity.

This also makes the LocalCompilationsActivity into a "singleTask"
activity, since it doesn't really make sense to have multiple instances of
this activity in the app's stack.

Bug: T172860
Change-Id: Ic7a932f411d2b09ee535719a54b859ab6832d6bf