Radically simplify ActionMode logic in PageActivity.
It turns out that there's no longer any need to differentiate between
ActionMode and SupportActionMode. (the Support library handles this
automatically under the hood)

- There's no further need for a custom "CompatActionMode" class.
- No further need to override both onActionModeStarted() and
  onSupportActionModeStarted().

This also fixes some possible crashes in older APIs, apparently because
we were sending a Bridge message to the WebView during initialization of
an ActionMode, which older WebViews don't seem to handle well. We were
doing this to get the currently-highlighted text as soon as the user
long-presses the WebView, and send it to eventlogging. However, we really
don't need the highlighted text at that point, since the event funnel will
continue with a "share" event (which will contain the highlighted text) or
an "abandon" event, both of which will have the install ID of the user for
tracking of the interaction.

Therefore, I'm simply removing the call to get the initial highlighted
text.

Bug: T173827
Change-Id: Ic0e88ddeb1a79b4e5fa65ada35151a5226356b4e