Fix inconsistent image loading on older APIs.
Aha, here's the real culprit:
We were setting progressiveRendering to true (without any clear reason),
and because we set it to true inside our face-detecting custom view, which
uses a postprocessor, it's possible that the postprocessing gets invoked
before the actual bitmap is fully loaded, causing it to fall into a funky
state.

Let's hope this fixes it definitively. I definitely can no longer
reproduce the issue on my devices.

Bug: T173896
Change-Id: I9a2c42d2f2f68a0c62c9790204c705b8710ac198