Update and generalize design of onboarding cards for the Feed.
(and specifically update the offline onboarding card.)

This removes the recently-added OnboardingCardView from the feed (which
was used for delivering the offline onboarding card), and instead
generalizes it to use the existing AnnouncementCardView, since the designs
are now identical.

An "Announcement" can now be constructed locally within the app (instead
of only from a content service response), and fed directly into an
announcement card in the feed.

This also introduces the idea of "local" app URLs, which are fake URLs
that can be plugged into the "action" of an announcement, and will cause
the affirmative button to take the user to the desired activity in the
app.

So, in addition to having static in-app onboarding cards, we'll now be
able to craft remote announcements that take the user to other specific
activities in the app, rather than only external links.

Bug: T172187
Bug: T140011
Change-Id: I2cb4d3e2caa19a42e236145fe0e01ec3d8532b5d