Remove the page title left behind when a page load fails
The page title that was set on the action when a page load fails
was cleared only when the 'Retry' button was clicked. Thus it was
left behind in lots of other cases like when the user switches to
other tabs etc., This isn't nice.

So, clear the page title whenever a page is loaded to ensure that
nothing like that happens. This also removes a callback method for
the 'Retry' button that just cleared the page title.

Bug: T163115
Bug: T173756
Change-Id: Id8a4c6f917c1c0a31a61151019388da1f80fae51