Don't destroy WebView explicitly.
Explicitly calling destroy() on the WebView seems to be having adverse
effects on the most recent version of the System WebView, whereby all
subsequent instances of the WebView can enter a bad state.

We started calling destroy() to resolve memory leaks that were found
previously, but basic functionality of the WebView obviously takes
priority. We'll need to re-audit memory leaks and fix any ones related to
the WebView in a different way (if they still exist).

Bug: T174401
Change-Id: I13e9967ceb6b76af978542c9482732b0b9bb351b