Update ZIM library to latest version.
This update to the zimdroid library features improvements in searching the
ZIM file's index by using the requested title or URL, plus the expected
namespace of the data item. ("namespace" in the sense of ZIM namespaces,
not Wikipedia namespaces.)

This fixes some edge cases that have been recently observed when trying to
fetch certain article titles, and also when trying to search in certain
character sets.

Bug: T175035
Bug: T175039
Change-Id: I3b0b9aced253f8e7452322fab16224845651492a