Add intent-filter for Zim files.
This allows the app to be launched (and land on the Offline Library screen)
when the user taps the DownloadManager notification after it has completed
downloading. It also allows the app to be launched if a Zim file is
clicked from an external application, e.g. a file manager app.

Bug: T172860
Change-Id: Ibafe8fa0c3a5c396b2361c312637949e2a4733ff