Clean up new setDesiredLocale
I was a bit quick on the draw merging If50de9da: Dmitry is right that
we should avoid recursing in setDesiredLocale.  With our current language
variant handling this shouldn't result in an infinite recursion, but it
could leave us open if things change, and in any case there's no real
benefit I can see to recursing here.

Change-Id: I313294085a489b560a4a585460247af9ecd07651