Fix fetching mobileview sections via XHR (viz. Chinese)
Shockingly, when making the XHR patch, I forgot to take into account that
the incoming response might be a `mobileview` response, versus a restbase
response. The two are slightly different, with a restbase response having
the "sections" object at the top level, while a mobileview response has
"sections" wrapped in a top level "mobileview" object.

Change-Id: I923646de6dbf247159a5ee79d1678d99f61dc6e4