Explicitly make offline pack available immediately upon finishing download.
The problem was that if the user goes to the Detail screen of a pack, then
starts downloading it, and *remains* on the detail screen until the pack
finishes downloading it, then the OfflineManager still doesn't actually
"know" about the new pack. (The OfflineManager only refreshes its
knowledge of existing packs when the user navigates back to the
LocalCompilations screen.)

This prevents the user from being able to click the "Main Page" button to
go to the main page of the compilation.

This patch explicitly pokes the newly-downloaded compilation into the
OfflineManager when we detect that the download has finished.

Bug: T175581
Change-Id: Ia4eab16e5ed6c19b65460e0bf291ac5e079b0ee5