Fix: Always ask for Write permission, not Read. (crash on Oreo)
When navigating to our LocalCompilations screen, we ask for permission to
READ external storage, so that it can scan the device for existing packs.

But the thing is, when the user wants to download a new pack, we actually
need the WRITE permission, so that the pack can be written to storage.
Older versions of Android don't really make the distinction between the
READ and WRITE permissions, but Oreo is actually starting to enforce this
distinction, and crashes if you try to write to the storage without the
explicit WRITE permission.

So, rather than asking for the WRITE permission separately (after having
already asked for the READ permission), let's just ask for the WRITE
permission in the first place, which will implicitly grant the READ
permission anyway, so that the user only has to deal with one permission
popup dialog.

Bug: T175580
Change-Id: Idc22de0c6c6208c70c05e9b26f1c62522c4aa436