Fix page loading issue on API 19.
Issue 2: When deserializing the remaining-sections response in JavaScript,
it looks like API 19 doesn't support responseType="json", along with
earlier APIs.  I could have sworn that this was supported on API 19 and
higher, but apparently it's only API 20 and higher.

In addition, I added a try-catch block around the json deserialization and
section rendering, so that exceptions thrown during this process don't
stop the whole JavaScript sequence from working, and instead proceed to
tell the Java layer that the page has finished loading.

Bug: T175928
Change-Id: If7759956fda6ec97b968b66cf1a55207a6ccd21a