Suppress redundant Fresco image caching
Currently, there are two subfolders in our net cache: okhttp_cache
(maintained by OkHttp) and image_cache (maintained by Fresco).

image_cache contains all of the images requested by Fresco, i.e., all
images requested for use in native Android components.

okhttp_cache also contains images, namely the images requested by the
WebView, which do not go through Fresco.

In a default setup, Fresco forbids outside caching of the images it
requests, since it maintains its own cache.  However, we use a customized
network fetcher (CacheableOkHttpNetworkFetcher) that allows OkHttp to
cache Fresco-requested images as well, so that they are available for use
in saved pages (this allows us to persist lead images, for example).
As a result, in our setup, okhttp_cache also contains copies of all images
in image_cache, in addition to the WebView requested images.

This makes Fresco's image_cache completely redundant, so we should disable
it, and save the user a substantial amount of disk space.

To give an example of the savings involved, loading all of the day's feed
content for yesterday resulted in 1.42 MB of data in cache before, and
1.01 MB in cache afterward.

Change-Id: I1d26802f00f5d4cfe389b87f460b0cf7c9bd5a25