Bump summary spec version requested to 1.2.0
This version introduces the extract_html field which (as the name
suggests) provides an HTML extract.  We don't need this, but it's used
by the mobile web, and if we request the same version that they do, it
should improve our cache hit rate thanks to their heavy traffic load.

Change-Id: Ib0c7554af5c3d636db74b1a14cee9e4442e239fd