Fix: add Accept-Language header to requests coming from the WebView.
We have a class called OkHttpWebViewClient which catches all network
requests coming from the WebView, and routes them through clients created
by OkHttpConnectionFactory.  This is good, because OkHttpConnectionFactory
automatically appends all the necessary header onto the request...

...*except* the "Accept-Language" header.  This header is only set in the
RetrofitFactory class, because it's dependent on the current Wiki to which
the request applies.

This patch updates OkHttpWebViewClient so that it *also* sets the
Accept-Language header. This way, any request made from within the WebView,
which may include an XHR to retrieve additional page content, will also
get the proper language header.

TODO: Find a way to set this header in one place.

Bug: T176374
Change-Id: I08a5ef2fc8b74106d9041195c788b8514f8e76bc