Open inter-language links and Gallery links in the same tab.
This patch makes PageActivity provide two separate newIntent..() builders,
for loading a requested page in a new tab or the current tab.

This now allows LangLinksActivity and GalleryActivity to load their pages
in the current tab.

And in other places in the code, this makes it more explicit whether the
requested page will be loaded in a new or current tab.

Bug: T176163
Bug: T156148
Change-Id: I64da49a9b49fb19fe62f3aa146e10dbfbb2ec74e