Fix: Prevent duplicate images in the gallery
Due to a mismatched comparison between normalized and non-normalized
filenames, images clicked to launch the gallery which contained spaces
in their filenames were appearing in the gallery twice (i.e., injected
a second time after not being found).

With this patch we inject underscores for spaces on both sides to ensure
we're comparing like with like.

Test cases (both en):
* Hester Thrale
* Gennady Golovkin

Change-Id: Iad727e9f06aa100208541ab35c8df12ede4b1e50