Never suppress the language selection UI
Our dev settings contain a setting to indicate whether the MediaWiki
installation the app is hitting supports a language code subdomain
component; for instance, a standard MW-Vagrant installation doesn't.

We shouldn't hide the language selection UI when this is set to false,
because the app language also affects the Accept-Language header sent,
which we may want to test even against a single-language wiki (for
testing language variants, for example.

Change-Id: Iada4a5863c532317df566ebca081c45fea2ada92