Fix bookmark fill-in logic when adding/removing from reading lists.
Our recent changes to pending-deletion of reading list pages has caused a
slight dent in the logic for filling in the bookmark icon in PageFragment
and the Featured Article card (which also controls the menu options that
appear for adding/deleting the page).  This fixes the behavior to be as it
should.

To reproduce the issue:
- Open an article, add it to a reading list via the bottom toolbar button,
  then tap the button again to remove it. Notice that the button is still
  filled in, and nothing happens when it's tapped.

Change-Id: I9f023a8e50e9a04f421e319ff373f71e3e3bed89