Fix/simplify RecyclerView item decorations (dividers).
This fixes (and simplifies) the way we draw the thin dividers between the
items in our RecyclerViews.  Some recent changes exposed a flaw in the
previous implementation, where it actually wasn't providing enough space
for the divider to be drawn (and instead drawing it on top of the item
itself.

Change-Id: Ie589ef5262561d90558a8309db0057f2e1e3ff9b