Fix broken logic for reloading page contents upon theme change.
One of our recent patches ( https://gerrit.wikimedia.org/r/379934 ) seems
to have broken the logic for reloading the page contents when coming back
from savedInstanceState.  This can be because of a theme change, or any
kind of event where the activity is paused and then reloaded.

For a quick example:
- Go to the feed, and load the featured article (or any article from the
  feed).
- Change the theme.
- Observe that the screen becomes blank.

Change-Id: I595bba915d409518c5c215c10894e8a4727dd929