Add on-scroll color transition for page toolbar icons
In light mode, updates the page toolbar icons to white when a lead image
is present, then transitions them to black over the course of the on-
scroll animation.

To facilitate this, the page toolbar search and tab buttons are moved out
of the options menu and incorporated directly into the toolbar, so that we
have access to their views and drawables consistently and as soon as the
toolbar is inflated, rather than depending on the vagaries of the options
menu state.

Icon color updating is incorporated into the existing flow for condition-
ally adjusting the toolbar's opacity.

Bug: T170304
Change-Id: Ib4e16380f657830ebc5f5ac575eaff46d986f060