Properly deserialize MwQueryResult, with postprocessing!
There are actually a couple of postprocessing steps that need to be
performed on MwQueryResult:

The juicy meat of the response is the "pages" array, which contains
the list of page properties that were requested. However, there is a
separate array called "redirects" that contains a list of redirects that
were performed by the server to come up with the list of titles. There may
also be a list called "converted" that contains a list of titles that were
converted from one language variant to another.

Previously, the postprocessing of redirects was performed as a manual
one-time function when sorting through search results.  And the
"converted" array was not handled at all.

This patch creates a special Gson type adapter that supports post-
processing, and allows MwQueryResult to be post-processed automatically,
so that no one can ever forget to resolve the redirects and the converted
titles.

Bug: T177657
Change-Id: Id8da4bfd6f84e8a10319097e59ec2bdb32ea28b5