Hygiene: Standardize MW API boolean argument format
Boolean parameters supplied to the MW API are considered to be true if
present in the query string and set to any value.  In other words,

* converttitles=
* converttitles=true
* converttitles=foo

all result in the API considering convertitles to be true.

The recommended format is the first in the list above;[1] so let's
standardize on that.

[1] https://www.mediawiki.org/wiki/API:Data_formats#Boolean_values

Change-Id: I0d406e50edb8585eb95bd9dae7a76341ef23f8ff