Fix collapsing of References.
The page library collapses tables that are contained in a <table> element.
However, the list of references at the bottom of articles is not actually
a table, so it needs to be manually wrapped in a <table> tag.

The iOS app actually does this in native code (?) before passing the HTML
to the WebView, but we can do it as part of our transforms.

Bug: T177863
Change-Id: I1212b8336635cf6d6742ff659c0794ea6ea93254