Android Gradle Plugin -> 3.0.0
Bumps the Android Gradle Plugin to 3.0.0 and makes the minimum required
changes to compile.  Primarily this involved adding a "default" flavor
dimension so that all product flavors could be assigned to a flavor
dimension as now required.[1]

Downgrades the new AppLinkUrlError lint issue to a warning to prevent
lint failures.  For more on this issue (and Firebase App Indexing, which
appears to help surface the app in Google Search results), see:

https://firebase.google.com/docs/app-indexing/

Additionally, I updated now-deprecated dependency configuration names to
their new counterparts.[2]

Please see the full Android Gradle Plugin 3.0.0 migration guide for
details.

https://developer.android.com/studio/build/gradle-plugin-3-0-0-migration.html

[1] https://developer.android.com/studio/build/gradle-plugin-3-0-0-migration.html#flavor_dimensions
[2] https://developer.android.com/studio/build/gradle-plugin-3-0-0-migration.html#new_configurations

Change-Id: I7df1b28de77a49d7104a260b1d1c2bb2ca72bca6