Improve properness of launching SavedPageSyncService on API 26.
Android Oreo (API 26) is much more harsh on apps that run in the
background, and requires us to enqueue background services in a more
delicate way.  For example, if an app is in the background, and it tries
to start a Service, it will actually crash.

This patch incorporates the new JobIntentService introduced in the Support
library v26.0, which correctly launches the service for API 26 and lower
APIs.

One of the side-effects of this patch is that we'll need to include the
WAKE_LOCK permission in the app from now on.

Change-Id: I216b1243cd6c74c64788cb5e8d45aa43e081b0b2