Overhaul bottom content.
This performs an overhaul and modernization of the bottom content that we
display in articles, as well as the way we present it.

I introduce a new component called LinearLayoutOverWebView, which can be
used as a generic container for components that need to scroll along with
WebView contents. This container can now be applied to both the lead image
and the bottom content.

As a side effect, this overhaul knocks out a number of long-standing
issues relating to the presentation of the bottom content and the lead
image view.

Bug: T143400
Bug: T176752
Bug: T157808
Bug: T122364
Bug: T120769
Bug: T102192
Bug: T96359
Bug: T88509
Change-Id: I4938c791df4a4d3db75a0684ec4d1178ef8d072c